chore(cleanup): dead code 제거 및 코드 가독성 개선
Resolve #146
