+++ src/app/_components/Header/component/ThemeToggleButton.tsx
@@ -2,8 +2,16 @@
 
 import ClickButton from '@/_components/ClickButton';
 import { Theme } from '@/_types/Header/Header.type';
-import ThemeToggleView from './ThemeToggleView';
+import ThemeIcon from '@/assets/Header/theme.svg';
 import useTheme from '@/_hooks/useTheme';
+
+const ThemeToggleView = () => {
+  return (
+    <span className="select-none border-gray-500 dark:border-gray-300 rounded-lg">
+      <ThemeIcon />
+    </span>
+  );
+};
 
 interface Props {
   theme: Theme;
