--- src/app/_styles/admin.css
@@ -1,6 +1,0 @@
-/* 최상단에 추가 */
-html,
-body {
-  height: 100%;
-  background-color: #e2e8f0;
-}
