--- src/app/dummyData/lecture-alarm/getLectureAlarm.json
@@ -1,21 +1,0 @@
-{
-  "data": {
-    "notices": [
-      {
-        "id": 1,
-        "title": "강의 공지 제목1",
-        "content": "강의 공지 내용1"
-      },
-      {
-        "id": 2,
-        "title": "강의 공지 제목2",
-        "content": "강의 공지 내용2"
-      },
-      {
-        "id": 3,
-        "title": "강의 공지 제목3",
-        "content": "강의 공지 내용3"
-      }
-    ]
-  }
-}
