--- src/app/dummyData/lecture/getContents.ts
@@ -1,20 +1,0 @@
-// {
-//     "title": "",
-//     "contents": "",
-//     "thumbnailId": "",
-//     "materialId": null,
-//     "maxCapacity": 0,
-//     "startTime": "",
-//     "endTime": "",
-//     "location": "",
-//     "category": "BACKEND",
-//     "speakerName": "",
-//     "speakerEmail": "",
-//     "speakerPosition": "",
-//     "speakerIntroduction": "",
-//     "speakerFilmography": [],
-//     "speakerPhotoId": ""
-// }
-
-export const contents =
-  'N4IgLgngDgpiBcIAmB7AxiANCNKB2YMBCA2qJLAiABYwCGSAlngOZYh1hgBOAzgqAA2MAG4xBCAMwBfbLgJEwpctDiJCADyXZNSxIFQawBrjAAkAxg4BwakNIC6slZURQ63OixdRq7eYWLwy4VSpddmDEAB0QQBdx4yNAFKbjAFUADQB6BIBJY0AUVsAE8cAPcZNABjrAD57jc0AFRsAA3sAGReNAE6HAFm7jQEAJwEqxwBk6wAHJ40ATpsALjsAHLsxjADFGFgBbOkAfTuNAHTXWwB1V7PzjQB2hwBKhwB1F40AbWsBU2eHAEFXAHQ7AAsXAAYXIwEVV4atMezUaZyQAIW56AGt7x6CYLRC-z0IDy1UAme2ADUGVnlhoAagdaOWMgBlFjaADqXjIBHlsAIuP1QCtQ4BLVcAE009AB0iSSxkAC2OATTnAD81xkAOrOAA6HjKjADAdgB2W2aATVXcYANVcABzXDOqAHAnjLDAAIdxkAIb2AXEGcYBWxcALqvGH4BBzPbhvD50b52dVPYI6IFURqABdHADiDxkAKD0nFqtQAGq8ZABhDxkAEePxQCUPYBUnsABh04wCu44BM9eJZMAELOAF56cjV6g0cl1jIBfIcAMTXGQAMrYAfduMTtdbsAI82AXQ69oBGQcAGC2AFtH3YAecaGqv1FCe1Be7y+aobfwBxoBiDquPpp0AuBPowqADB7jLzFSq6oAJyeaeWMgAaa2mAHm6cYAZzsAJy0u9240MRKy2B4GqhOFxuOgeNuBRC0BjMNjYTg8fjwISicRSWQ4fA+PT+dvqCaXbAkYxiAJVdgAUMwe9bXiAKBajAHxIAAMowvDaBwXB8AIIDoc4egAIxft4ijKMeiCCGhYDpIQExeD+pF+L8jjOK47ieHIDG+P+sFGuAJqIMkaSZLkeSRIAAe0APzGAAFBBkSAAgT9IVMYgCoE4ANZ0AJTQYezEgJR6E0TAdGcQo3F6aebEXhx35mX+el8aEIDJFSdKMiyomspyOkwRqBnUbR9F2WRAEgJZ56XqZv4hbxQH8d2IDwoigCIk4AN0PQjJAAiMC8OMeDGAAKtQzCfPemkAAVlIAKWM+bp5FhaxEWeL5jb0EwrDsE+2Gvvp74SPAMhRYxPEao5AkgIAKbOADkzxiAADNgCDA9BR6hfBSCITAKFUZ1WEvqAeHcIRxFcfZ9X+UZJm2dFTH1eF7FBVdI2GnFTlCRk0LGIAvTWABur0SAGOjnRrsYJRlOUOKAAw9gANYwKxiAG9DgAC4+KgAx7T5y2wWdgVDeZN2NXdmMnaFY0JZ531-QDwyvZk0qAD6jdqACOTgCh44KsMI7CyM2OzqN+VR533cNFk49ZvNYwTz3jWMkx0MYgCYNQjnmACxLOKABkzgAy48YgCIE3y8Ss7V7NAA';
