--- src/app/dummyData/lecture/getLectureContents.json
@@ -1,129 +1,0 @@
-{
-  "type": "doc",
-  "content": [
-    {
-      "type": "heading",
-      "attrs": { "level": 3 },
-      "content": [{ "type": "text", "text": "강의 소개" }]
-    },
-    {
-      "type": "paragraph",
-      "content": [
-        {
-          "type": "text",
-          "marks": [{ "type": "textStyle", "attrs": { "color": "" } }],
-          "text": "\"이 강의는 UX/UI 디자인의 기본 개념부터 실무 적용까지 다루며, Figma를 활용한 디자인 시스템 구축, 프로토타이핑, "
-        },
-        { "type": "hardBreak", "marks": [{ "type": "textStyle", "attrs": { "color": "" } }] },
-        {
-          "type": "text",
-          "marks": [{ "type": "textStyle", "attrs": { "color": "" } }],
-          "text": "인터랙션 디자인, 사용자 테스트 등을 실습합니다. UX 원칙과 최신 트렌드를 학습하고, 실제 사례 분석을 통해 "
-        },
-        { "type": "hardBreak", "marks": [{ "type": "textStyle", "attrs": { "color": "" } }] },
-        {
-          "type": "text",
-          "marks": [{ "type": "textStyle", "attrs": { "color": "" } }],
-          "text": "실무에서 바로 적용할 수 있는 방법론을 익힙니다. 초보자부터 실무자까지 쉽게 따라 할 수 있도록 구성되어 있으며, "
-        },
-        { "type": "hardBreak", "marks": [{ "type": "textStyle", "attrs": { "color": "" } }] },
-        {
-          "type": "text",
-          "marks": [{ "type": "textStyle", "attrs": { "color": "" } }],
-          "text": "실습과 프로젝트 기반 학습을 통해 실질적인 결과물을 만들 수 있습니다.\""
-        }
-      ]
-    },
-    { "type": "paragraph" },
-    {
-      "type": "heading",
-      "attrs": { "level": 3 },
-      "content": [
-        {
-          "type": "text",
-          "marks": [{ "type": "textStyle", "attrs": { "color": "" } }],
-          "text": "강의 목차"
-        }
-      ]
-    },
-    {
-      "type": "orderedList",
-      "attrs": { "start": 1 },
-      "content": [
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [{ "type": "text", "text": "UX/UI 디자인이란? (차이점과 개념 정리)" }]
-            }
-          ]
-        },
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [{ "type": "text", "text": "UX 원칙과 최신 디자인 트렌드" }]
-            }
-          ]
-        },
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [{ "type": "text", "text": "사용자 중심 디자인(Design Thinking)  개요" }]
-            }
-          ]
-        }
-      ]
-    },
-    { "type": "paragraph" },
-    {
-      "type": "heading",
-      "attrs": { "level": 3 },
-      "content": [{ "type": "text", "text": "추천 대상" }]
-    },
-    {
-      "type": "orderedList",
-      "attrs": { "start": 1 },
-      "content": [
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [
-                {
-                  "type": "text",
-                  "text": "UX/UI 디자인 경험이 없지만 기본 개념을 배우고 싶은 사람"
-                }
-              ]
-            }
-          ]
-        },
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [
-                { "type": "text", "text": "디자인 경험이 없지만, UX/UI 분야로 진입하고 싶은 사람" }
-              ]
-            }
-          ]
-        },
-        {
-          "type": "listItem",
-          "content": [
-            {
-              "type": "paragraph",
-              "content": [{ "type": "text", "text": "Figma 같은 디자인 툴을 처음 접하는 사람" }]
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
