--- src/app/dummyData/lecture/getQuestion.json
@@ -1,34 +1,0 @@
-{
-  "data": [
-    {
-      "id": 1,
-      "user_id": 222,
-      "lecture_id": 1,
-      "contents": "222 유저가 강의 1에 한 사전질문"
-    },
-    {
-      "id": 2,
-      "user_id": 222,
-      "lecture_id": 2,
-      "contents": "222 유저가 강의 2에 한 사전질문"
-    },
-    {
-      "id": 3,
-      "user_id": 444,
-      "lecture_id": 1,
-      "contents": "444 유저가 강의 1에 한 사전질문"
-    },
-    {
-      "id": 4,
-      "user_id": 444,
-      "lecture_id": 2,
-      "contents": "444 유저가 강의 2에 한 사전질문"
-    },
-    {
-      "id": 5,
-      "user_id": 666,
-      "lecture_id": 1,
-      "contents": "666 유저가 강의 1에 한 사전질문"
-    }
-  ]
-}
