--- src/app/dummyData/lecture/getReview.json
@@ -1,31 +1,0 @@
-{
-  "data": {
-    "lectureId": 1,
-    "reviews": [
-      {
-        "id": 1,
-        "userId": 222,
-        "username": "이이이",
-        "content": "222 유저의 리뷰",
-        "rating": 5.0,
-        "createdAt": "2025-03-07T11:30:00"
-      },
-      {
-        "id": 2,
-        "userId": 444,
-        "username": "사사사",
-        "content": "444 유저의 리뷰",
-        "rating": 4.0,
-        "createdAt": "2025-03-07T12:30:00"
-      },
-      {
-        "id": 3,
-        "userId": 666,
-        "username": "육육육",
-        "content": "666 유저의 리뷰",
-        "rating": 5.0,
-        "createdAt": "2025-03-07T13:30:00"
-      }
-    ]
-  }
-}
