--- src/app/dummyData/main/maindto.json
@@ -1,5 +1,0 @@
-{
-  "data": {
-    "user": "data"
-  }
-}
