--- src/app/dummyData/timetable/getTimeWish.json
@@ -1,27 +1,0 @@
-{
-  "responses": [
-    {
-      "lectureId": 1,
-      "category": "BACKEND",
-      "thumbnailUri": "uploads/lecture/thumbnail/123456789.jpg",
-      "title": "공간지능 혁신을 통한 온오프라인 통합 경험의 미래",
-      "contents": "Part 1: 온오프라인 경험을 연결하는 네이버 지도가 공간지능과 만나 제공하게 될 미래 모습을 소개합니다.",
-      "speakerName": "Cole Palmer",
-      "startTime": "2025-03-01T10:00:00",
-      "endTime": "2025-03-01T12:00:00",
-      "location": "온라인"
-    },
-    {
-      "lectureId": 2,
-      "category": "BACKEND",
-      "thumbnailUri": "uploads/lecture/thumbnail/123456789.jpg",
-      "title": "공간지능 혁신을 통한 온오프라인 통합 경험의 미래",
-      "contents": "Part 1: 온오프라인 경험을 연결하는 네이버 지도가 공간지능과 만나 제공하게 될 미래 모습을 소개합니다.",
-      "speakerName": "Cole Palmer",
-      "startTime": "2025-03-01T10:00:00",
-      "endTime": "2025-03-01T12:00:00",
-      "location": "온라인"
-    }
-  ],
-  "total": 2
-}
