--- src/app/dummyData/timetable/getTimetable.json
@@ -1,67 +1,0 @@
-{
-  "reservation": [
-    {
-      "reservationId": 1,
-      "userId": 222,
-      "lectureId": 1,
-      "title": "강의 제목1",
-      "startTime": "2025-03-04T10:00:00",
-      "endTime": "2025-03-04T12:00:00",
-      "currentReservation": 12,
-      "capacity": 45,
-      "location": "101호",
-      "speakerName": "황효주"
-    },
-    {
-      "reservationId": 2,
-      "userId": 222,
-      "lectureId": 2,
-      "title": "강의 제목2",
-      "startTime": "2025-03-04T14:00:00",
-      "endTime": "2025-03-04T16:00:00",
-      "currentReservation": 5,
-      "capacity": 80,
-      "location": "102호",
-      "speakerName": "홍길동"
-    },
-    {
-      "reservationId": 3,
-      "userId": 222,
-      "lectureId": 3,
-      "title": "강의 제목3",
-      "startTime": "2025-03-05T09:00:00",
-      "endTime": "2025-03-05T11:00:00",
-      "currentReservation": 7,
-      "capacity": 20,
-      "location": "103호",
-      "speakerName": "고길동"
-    }
-  ],
-
-  "wishlist": [
-    {
-      "wishlistId": 1,
-      "userId": 222,
-      "lectureId": 4,
-      "title": "강의 제목4",
-      "startTime": "2025-03-04T14:00:00",
-      "endTime": "2025-03-04T16:00:00",
-      "currentReservation": 12,
-      "capacity": 45,
-      "location": "104호",
-      "speakerName": "신짱구"
-    },
-    {
-      "wishlistId": 2,
-      "userId": 222,
-      "lectureId": 5,
-      "title": "강의 제목5",
-      "startTime": "2025-03-05T14:00:00",
-      "endTime": "2025-03-05T16:00:00",
-      "currentReservation": 12,
-      "capacity": 45,
-      "location": "105호",
-      "speakerName": "김철수"
-    }
-  ]
-}
