--- src/app/dummyData/timetable/getTimetableRecommand.json
@@ -1,24 +1,0 @@
-{
-  "lectures": [
-    {
-      "id": 1,
-      "title": "추천강의제목1",
-      "category": "카테고리1",
-      "startTime": "2025-03-05T10:00:00",
-      "endTime": "2025-03-05T11:00:00",
-      "speakerName": "홍길동",
-      "speakerImageUri": "",
-      "location": "101호"
-    },
-    {
-      "id": 2,
-      "title": "추천강의제목2",
-      "category": "카테고리2",
-      "startTime": "2025-03-05T10:00:00",
-      "endTime": "2025-03-05T11:00:00",
-      "speakerName": "홍길동",
-      "speakerImageUri": "",
-      "location": "102호"
-    }
-  ]
-}
